requirements.txt
click==7.1.2
Flask==1.1.2
grpcio==1.30.0
grpcio-tools==1.30.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.21
pbkdf2==1.3
protobuf==3.12.2
six==1.15.0
Werkzeug==1.0.1
